Integrated Transparent over Wall (ToW)(Thanks Jun Lee, PunkDM) Including components, -pdnsd, redsocks2, shadowsocks, sshpass, obfuscation ssh -Scripts for ToW -Add chains to firewall NAT table -Web GUI for ToW setting, under Advanced/ToW Proxy -Also you can start up ToW manually by run: service tow start/stop/restart -And start individual service by run: service pdnsd/redsocks2/shadowsocks/sshpass/sshofc start/stop/restart -All of generated scripts are in directory /etc/tow 2014.1.6.

+++ release/src/btools/libfoo.pl
@@ -319,6 +319,8 @@ sub fixDyn
 	fixDynDep("xxsubtype.so", "libpython2.7.so.1.0");
 	fixDynDep("zlibmodule.so", "libpython2.7.so.1.0");
 	fixDynDep("zlibmodule.so", "libz.so.1");
+#tow, bwq518
+	fixDynDep("ssh_ofc", "libresolv.so.0");
 #iptables, bwq518
 	fixDynDep("xtables-multi", "libxtables.so.10.0.0");
 	fixDynDep("xtables-multi", "libiptc.so.0.0.0");
@@ -589,8 +591,12 @@ sub genSO
 	}
 #	print "$cmd -u... ${arc}\n";	
 	if (scalar(@used) == 0) {
-		print "$name: WARNING: Library is not used by anything, deleting...\n";
-		unlink $so;
+		if (($name ne 'libresolv.so.0') && ($name ne 'libnsl.so.0'))
+		{
+			print "$name: WARNING: Library is not used by anything, deleting...\n";
+			unlink $so;
+		}
+
 		@so_fname=split(/\./,$name);
 		$so_symlinks=dirname($so)."\/".@so_fname[0]."\.so\*";
 		print "$so_symlinks: WARNING: Symbol links is invalid, deleting...\n";
